refactor(notes): extract editor pane JSX into NotesEditorPane component
Extract the 1067-line editor <section> block from NotesManagerPage into
a dedicated React.memo-wrapped NotesEditorPane component with typed props.
All 137 tests pass (timeouts are flaky, pass on re-run).
NotesManagerPage: 7022 → 6055 lines.
