+++ apps/packages/ui/src/components/Notes/NotesManagerPage.tsx
@@ -34,6 +34,7 @@ import { normalizeChatRole } from "@/utils/normalize-chat-role"
 import { useScrollToServerCard } from "@/hooks/useScrollToServerCard"
 import { MarkdownPreview } from "@/components/Common/MarkdownPreview"
 import NotesEditorHeader from "@/components/Notes/NotesEditorHeader"
+import NotesEditorPane from "@/components/Notes/NotesEditorPane"
 import NotesSidebar from "@/components/Notes/NotesSidebar"
 import NotesListPanel from "@/components/Notes/NotesListPanel"
 import NotesGraphModal from "@/components/Notes/NotesGraphModal"
@@ -5512,1073 +5513,105 @@ const NotesManagerPage: React.FC = () => {
       )}
 
       {/* Editor Panel */}
-      <section
-        id={NOTES_EDITOR_REGION_ID}
-        tabIndex={-1}
-        role="region"
-        aria-label={t('option:notesSearch.editorRegionLabel', {
-          defaultValue: 'Note editor'
-        })}
-        aria-describedby={NOTES_SHORTCUTS_SUMMARY_ID}
-        aria-busy={loadingDetail}
-        data-testid="notes-editor-region"
-        className={`flex-1 flex flex-col overflow-hidden rounded-lg border border-border bg-surface ${
-          isMobileViewport ? 'ml-0' : 'ml-4'
-        }`}
-        aria-disabled={editorDisabled}
-      >
-        {isMobileViewport && (
-          <div className="border-b border-border px-4 py-2">
-            <Button
-              size="large"
-              onClick={() => setMobileSidebarOpen(true)}
-              data-testid="notes-mobile-open-list-button"
-              className="min-h-[44px]"
-            >
-              {t('option:notesSearch.openMobileSidebar', {
-                defaultValue: 'Browse notes'
-              })}
-            </Button>
-          </div>
-        )}
-        <NotesEditorHeader
-          title={title}
-          selectedId={selectedId}
-          backlinkConversationId={backlinkConversationId}
-          backlinkConversationLabel={backlinkConversationLabel}
-          backlinkMessageId={backlinkMessageId}
-          sourceLinks={noteRelations.sources}
-          editorDisabled={editorDisabled}
-          openingLinkedChat={openingLinkedChat}
-          editorMode={editorMode}
-          hasContent={content.trim().length > 0}
-          canSave={
-            !editorDisabled &&
-            (title.trim().length > 0 || content.trim().length > 0)
-          }
-          canGenerateFlashcards={!editorDisabled && content.trim().length > 0}
-          canExport={Boolean(title || content)}
-          canDuplicate={!editorDisabled && (title.trim().length > 0 || content.trim().length > 0)}
-          canPin={!editorDisabled && selectedId != null}
-          isPinned={selectedNotePinned}
-          templateOptions={NOTE_TEMPLATES.map((template) => ({
-            id: template.id,
-            label: template.label
-          }))}
-          isSaving={saving}
-          canDelete={!editorDisabled && isOnline && selectedId != null}
-          isDirty={isDirty}
-          onOpenLinkedConversation={() => {
-            void openLinkedConversation()
-          }}
-          onOpenSourceLink={(sourceId, sourceLabel) => {
-            openLinkedSource(sourceId, sourceLabel)
-          }}
-          onNewNote={() => {
-            void handleNewNote()
-          }}
-          onApplyTemplate={(templateId) => {
-            void handleNewNote(templateId)
-          }}
-          onDuplicate={() => {
-            void duplicateSelectedNote()
-          }}
-          onTogglePin={() => {
-            if (selectedId == null) return
-            void toggleNotePinned(selectedId)
-          }}
-          onChangeEditorMode={(nextMode) => {
-            setEditorMode(nextMode)
-          }}
-          onCopy={(mode) => {
-            void copySelected(mode)
-          }}
-          onGenerateFlashcards={handleGenerateFlashcardsFromNote}
-          onExport={(format) => {
-            exportSelected(format)
-          }}
-          onSave={() => {
-            void saveNote()
-          }}
-          onDelete={() => {
-            void deleteNote()
-          }}
-        />
-        <div className="flex-1 flex flex-col px-4 py-3 overflow-auto">
-          {loadingDetail && (
-            <div
-              className="mb-3 inline-flex w-fit items-center gap-2 rounded border border-border bg-surface2 px-3 py-1.5 text-[12px] text-text-muted"
-              role="status"
-              aria-live="polite"
-              data-testid="notes-editor-loading-detail"
-            >
-              <Spin size="small" />
-              <span>
-                {t('option:notesSearch.loadingDetail', {
-                  defaultValue: 'Loading note details...'
-                })}
-              </span>
-            </div>
-          )}
-          <div className="flex items-center gap-2">
-            <Input
-              placeholder={t('option:notesSearch.titlePlaceholder', {
-                defaultValue: 'Title'
-              })}
-              value={title}
-              onChange={(e) => {
-                setTitle(e.target.value)
-                setIsDirty(true)
-                setSaveIndicator('dirty')
-                setMonitoringNotice(null)
-                markManualEdit()
-              }}
-              disabled={editorDisabled}
-              ref={titleInputRef}
-              className="bg-transparent hover:bg-surface2 focus:bg-surface2 transition-colors"
-            />
-            <Tooltip
-              title={t('option:notesSearch.generateTitleTooltip', {
-                defaultValue: 'Generate title from content'
-              })}
-            >
-              <Button
-                size="small"
-                onClick={() => {
-                  void suggestTitle()
-                }}
-                disabled={editorDisabled || !isOnline || content.trim().length === 0}
-                loading={titleSuggestionLoading}
-                icon={(<SparklesIcon className="w-4 h-4" />) as any}
-                aria-label={t('option:notesSearch.generateTitleTooltip', {
-                  defaultValue: 'Generate title from content'
-                })}
-                data-testid="notes-generate-title-button"
-              >
-                {t('option:notesSearch.generateTitleAction', {
-                  defaultValue: 'Generate title'
-                })}
-              </Button>
-            </Tooltip>
-            {canSwitchTitleStrategy ? (
-              <Select
-                size="small"
-                className="min-w-[170px]"
-                value={effectiveTitleSuggestStrategy}
-                options={titleStrategyOptions}
-                onChange={(value) => {
-                  const normalized = normalizeNotesTitleStrategy(value)
-                  if (!normalized) return
-                  setTitleSuggestStrategy(normalized)
-                  void setSetting(NOTES_TITLE_SUGGEST_STRATEGY_SETTING, normalized)
-                }}
-                disabled={editorDisabled || !isOnline || titleSuggestionLoading}
-                aria-label={t('option:notesSearch.titleStrategyLabel', {
-                  defaultValue: 'Title generation strategy'
-                })}
-                data-testid="notes-title-strategy-select"
-              />
-            ) : null}
-            <Tooltip
-              title={t('option:notesSearch.shortcutHelpTooltip', {
-                defaultValue: 'Keyboard shortcuts'
-              })}
-            >
-              <Button
-                size="small"
-                type="text"
-                onClick={() => setShortcutHelpOpen(true)}
-                aria-label={t('option:notesSearch.shortcutHelpTooltip', {
-                  defaultValue: 'Keyboard shortcuts'
-                })}
-                data-testid="notes-shortcuts-help-button"
-              >
-                {t('option:notesSearch.shortcutHelpLabel', {
-                  defaultValue: 'Keyboard shortcuts'
-                })}
-              </Button>
-            </Tooltip>
-          </div>
-          <div className="mt-3">
-            <Select
-              mode="tags"
-              allowClear
-              placeholder={t('option:notesSearch.keywordsEditorPlaceholder', {
-                defaultValue: 'Keywords (tags)'
-              })}
-              data-testid="notes-keywords-editor"
-              className="w-full"
-              value={editorKeywords}
-              onSearch={(txt) => {
-                if (isOnline) void debouncedLoadKeywordSuggestions(txt)
-              }}
-              onChange={(vals) => {
-                setEditorKeywords(vals as string[])
-                setIsDirty(true)
-                setSaveIndicator('dirty')
-                setMonitoringNotice(null)
-                markManualEdit()
-              }}
-              options={keywordOptions.map((keyword) => ({
-                label: renderKeywordLabelWithFrequency(keyword, {
-                  includeCount: true,
-                  testIdPrefix: 'notes-keyword-editor-option-label'
-                }),
-                value: keyword
-              }))}
-              disabled={editorDisabled}
-            />
-            <Typography.Text
-              type="secondary"
-              className="block text-[11px] mt-1 text-text-muted"
-            >
-              {t('option:notesSearch.tagsHelp', {
-                defaultValue:
-                  'Keywords help you find this note using the keyword filter on the left.'
-              })}
-            </Typography.Text>
-            {saveIndicatorText && (
-              <Typography.Text
-                type={saveIndicator === 'error' ? 'danger' : 'secondary'}
-                className="block text-[11px] mt-1 text-text-muted"
-                aria-live="polite"
-              >
-                {saveIndicatorText}
-              </Typography.Text>
-            )}
-            {offlineStatusText && (
-              <Typography.Text
-                type={currentOfflineDraft?.syncState === 'conflict' ? 'danger' : 'secondary'}
-                className="block text-[11px] mt-1 text-text-muted"
-                aria-live="polite"
-                data-testid="notes-offline-sync-status"
-              >
-                {offlineStatusText}
-              </Typography.Text>
-            )}
-            {remoteVersionInfo && (
-              <div
-                className="mt-2 rounded border border-warn/50 bg-warn/10 px-2 py-1 text-[12px] text-warn"
-                role="status"
-                data-testid="notes-stale-version-warning"
-              >
-                <span>
-                  {t('option:notesSearch.staleVersionWarning', {
-                    defaultValue:
-                      'A newer version is available on the server (v{{version}}).',
-                    version: remoteVersionInfo.version
-                  })}
-                </span>
-                <Button
-                  type="link"
-                  size="small"
-                  className="!px-1"
-                  onClick={() => {
-                    if (selectedId == null) return
-                    void handleSelectNote(selectedId)
-                  }}
-                  data-testid="notes-stale-version-reload"
-                >
-                  {t('option:notesSearch.reloadNoteAction', {
-                    defaultValue: 'Reload note'
-                  })}
-                </Button>
-              </div>
-            )}
-            {monitoringNotice && (
-              <div
-                className={`mt-2 rounded border px-2 py-2 text-[12px] ${monitoringNoticeClasses}`}
-                role="alert"
-                aria-live="polite"
-                data-testid="notes-monitoring-alert"
-              >
-                <div className="font-medium">{monitoringNotice.title}</div>
-                <div className="mt-1">{monitoringNotice.guidance}</div>
-              </div>
-            )}
-          </div>
-          {selectedId != null && (
-            <div
-              className="mt-3 grid grid-cols-1 gap-3 xl:grid-cols-2"
-              data-testid="notes-graph-relation-panels"
-            >
-              <div className="rounded-lg border border-border bg-surface2 p-3">
-                <Typography.Text
-                  className="text-[11px] uppercase tracking-[0.08em] text-text-muted"
-                  data-testid="notes-related-heading"
-                >
-                  {t('option:notesSearch.relatedNotesHeading', {
-                    defaultValue: 'Related notes'
-                  })}
-                </Typography.Text>
-                <Button
-                  size="small"
-                  className="mt-2"
-                  onClick={openGraphModal}
-                  data-testid="notes-open-graph-view"
-                >
-                  {t('option:notesSearch.graphOpenButton', {
-                    defaultValue: 'Open graph view'
-                  })}
-                </Button>
-                <div className="mt-2 flex items-center gap-2">
-                  <select
-                    className="flex-1 rounded border border-border bg-surface px-2 py-1 text-sm text-text"
-                    value={manualLinkTargetId ?? ''}
-                    onChange={(event) => {
-                      const value = event.target.value
-                      setManualLinkTargetId(value || null)
-                    }}
-                    disabled={manualLinkSaving || editorDisabled || manualLinkOptions.length === 0}
-                    data-testid="notes-manual-link-target-select"
-                  >
-                    <option value="">
-                      {t('option:notesSearch.manualLinkTargetPlaceholder', {
-                        defaultValue: 'Select a note to link'
-                      })}
-                    </option>
-                    {manualLinkOptions.map((option) => (
-                      <option key={option.value} value={option.value}>
-                        {option.label}
-                      </option>
-                    ))}
-                  </select>
-                  <Button
-                    size="small"
-                    type="primary"
-                    onClick={() => {
-                      void createManualLink()
-                    }}
-                    disabled={!manualLinkTargetId || editorDisabled}
-                    loading={manualLinkSaving}
-                    data-testid="notes-manual-link-add"
-                  >
-                    {t('option:notesSearch.manualLinkAdd', {
-                      defaultValue: 'Add link'
-                    })}
-                  </Button>
-                </div>
-                <Typography.Text
-                  type="secondary"
-                  className="block mt-2 text-[11px] text-text-muted"
-                >
-                  {t('option:notesSearch.manualLinksHeading', {
-                    defaultValue: 'Manual links'
-                  })}
-                </Typography.Text>
-                {noteRelations.manualLinks.length === 0 ? (
-                  <Typography.Text
-                    type="secondary"
-                    className="block mt-1 text-[12px] text-text-muted"
-                    data-testid="notes-manual-links-empty"
-                  >
-                    {t('option:notesSearch.manualLinksEmpty', {
-                      defaultValue: 'No manual links yet.'
-                    })}
-                  </Typography.Text>
-                ) : (
-                  <div className="mt-1 flex flex-wrap gap-1.5" data-testid="notes-manual-links-list">
-                    {noteRelations.manualLinks.map((link) => (
-                      <div
-                        key={link.edgeId}
-                        className="inline-flex items-center gap-1 rounded border border-border bg-surface px-2 py-1"
-                      >
-                        <button
-                          type="button"
-                          className="text-xs text-text hover:underline"
-                          onClick={() => {
-                            void handleSelectNote(link.noteId)
-                          }}
-                        >
-                          {link.title}
-                        </button>
-                        <button
-                          type="button"
-                          className="text-xs text-danger hover:underline"
-                          onClick={() => {
-                            void removeManualLink(link.edgeId)
-                          }}
-                          disabled={manualLinkDeletingEdgeId === link.edgeId}
-                          aria-label={t('option:notesSearch.manualLinkRemoveAria', {
-                            defaultValue: `Remove manual link ${link.title}`
-                          })}
-                          data-testid={`notes-manual-link-remove-${link.edgeId.replace(/[^a-z0-9_-]/gi, '_')}`}
-                        >
-                          {manualLinkDeletingEdgeId === link.edgeId
-                            ? t('option:notesSearch.manualLinkRemoving', {
-                                defaultValue: 'Removing...'
-                              })
-                            : t('option:notesSearch.manualLinkRemove', {
-                                defaultValue: 'Remove'
-                              })}
-                        </button>
-                      </div>
-                    ))}
-                  </div>
-                )}
-                {noteNeighborsLoading ? (
-                  <Typography.Text
-                    type="secondary"
-                    className="block mt-2 text-[12px] text-text-muted"
-                  >
-                    {t('option:notesSearch.relatedNotesLoading', {
-                      defaultValue: 'Loading related notes...'
-                    })}
-                  </Typography.Text>
-                ) : noteNeighborsError ? (
-                  <Typography.Text
-                    type="danger"
-                    className="block mt-2 text-[12px]"
-                    data-testid="notes-related-error"
-                  >
-                    {t('option:notesSearch.relatedNotesError', {
-                      defaultValue: 'Could not load related notes.'
-                    })}
-                  </Typography.Text>
-                ) : noteRelations.related.length === 0 ? (
-                  <Typography.Text
-                    type="secondary"
-                    className="block mt-2 text-[12px] text-text-muted"
-                    data-testid="notes-related-empty"
-                  >
-                    {t('option:notesSearch.relatedNotesEmpty', {
-                      defaultValue: 'No related notes yet.'
-                    })}
-                  </Typography.Text>
-                ) : (
-                  <div className="mt-2 flex flex-wrap gap-1.5" data-testid="notes-related-list">
-                    {noteRelations.related.map((note) => (
-                      <button
-                        key={`related-${note.id}`}
-                        type="button"
-                        className="rounded border border-border bg-surface px-2 py-1 text-left text-xs text-text hover:bg-surface3"
-                        onClick={() => {
-                          void handleSelectNote(note.id)
-                        }}
-                      >
-                        {note.title}
-                      </button>
-                    ))}
-                  </div>
-                )}
-              </div>
-              <div className="rounded-lg border border-border bg-surface2 p-3">
-                <Typography.Text
-                  className="text-[11px] uppercase tracking-[0.08em] text-text-muted"
-                  data-testid="notes-backlinks-heading"
-                >
-                  {t('option:notesSearch.backlinksHeading', {
-                    defaultValue: 'Backlinks'
-                  })}
-                </Typography.Text>
-                {noteNeighborsLoading ? (
-                  <Typography.Text
-                    type="secondary"
-                    className="block mt-2 text-[12px] text-text-muted"
-                  >
-                    {t('option:notesSearch.backlinksLoading', {
-                      defaultValue: 'Loading backlinks...'
-                    })}
-                  </Typography.Text>
-                ) : noteNeighborsError ? (
-                  <Typography.Text
-                    type="danger"
-                    className="block mt-2 text-[12px]"
-                    data-testid="notes-backlinks-error"
-                  >
-                    {t('option:notesSearch.backlinksError', {
-                      defaultValue: 'Could not load backlinks.'
-                    })}
-                  </Typography.Text>
-                ) : noteRelations.backlinks.length === 0 ? (
-                  <Typography.Text
-                    type="secondary"
-                    className="block mt-2 text-[12px] text-text-muted"
-                    data-testid="notes-backlinks-empty"
-                  >
-                    {t('option:notesSearch.backlinksEmpty', {
-                      defaultValue: 'No backlinks yet.'
-                    })}
-                  </Typography.Text>
-                ) : (
-                  <div className="mt-2 flex flex-wrap gap-1.5" data-testid="notes-backlinks-list">
-                    {noteRelations.backlinks.map((note) => (
-                      <button
-                        key={`backlink-${note.id}`}
-                        type="button"
-                        className="rounded border border-border bg-surface px-2 py-1 text-left text-xs text-text hover:bg-surface3"
-                        onClick={() => {
-                          void handleSelectNote(note.id)
-                        }}
-                      >
-                        {note.title}
-                      </button>
-                    ))}
-                  </div>
-                )}
-              </div>
-            </div>
-          )}
-          {editorMode !== 'preview' && (
-            <div className="mt-3 flex items-center flex-wrap gap-1 rounded-lg border border-border bg-surface2 p-2">
-              <div
-                className="mr-2 inline-flex items-center gap-1 rounded-md border border-border bg-surface px-1 py-0.5"
-                role="group"
-                aria-label={t('option:notesSearch.inputModeGroup', {
-                  defaultValue: 'Input mode'
-                })}
-                data-testid="notes-input-mode-toggle"
-              >
-                <Button
-                  size="small"
-                  type={editorInputMode === 'markdown' ? 'primary' : 'text'}
-                  onClick={() => handleEditorInputModeChange('markdown')}
-                  disabled={editorDisabled}
-                  data-testid="notes-input-mode-markdown"
-                >
-                  {t('option:notesSearch.inputModeMarkdown', {
-                    defaultValue: 'Markdown'
-                  })}
-                </Button>
-                <Button
-                  size="small"
-                  type={editorInputMode === 'wysiwyg' ? 'primary' : 'text'}
-                  onClick={() => handleEditorInputModeChange('wysiwyg')}
-                  disabled={editorDisabled}
-                  data-testid="notes-input-mode-wysiwyg"
-                >
-                  {t('option:notesSearch.inputModeWysiwyg', {
-                    defaultValue: 'WYSIWYG'
-                  })}
-                </Button>
-              </div>
-              <Typography.Text
-                type="secondary"
-                className="text-[11px] mr-1 uppercase tracking-[0.08em]"
-              >
-                {t('option:notesSearch.formattingLabel', {
-                  defaultValue: 'Formatting'
-                })}
-              </Typography.Text>
-              <Tooltip title={t('option:notesSearch.toolbarBoldTooltip', { defaultValue: 'Bold' })}>
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<BoldIcon className="w-4 h-4" />) as any}
-                  onClick={() => applyMarkdownToolbarAction('bold')}
-                  disabled={editorDisabled}
-                  aria-label={t('option:notesSearch.toolbarBoldTooltip', { defaultValue: 'Bold' })}
-                  data-testid="notes-toolbar-bold"
-                />
-              </Tooltip>
-              <Tooltip title={t('option:notesSearch.toolbarItalicTooltip', { defaultValue: 'Italic' })}>
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<ItalicIcon className="w-4 h-4" />) as any}
-                  onClick={() => applyMarkdownToolbarAction('italic')}
-                  disabled={editorDisabled}
-                  aria-label={t('option:notesSearch.toolbarItalicTooltip', { defaultValue: 'Italic' })}
-                  data-testid="notes-toolbar-italic"
-                />
-              </Tooltip>
-              <Tooltip title={t('option:notesSearch.toolbarHeadingTooltip', { defaultValue: 'Heading' })}>
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<HeadingIcon className="w-4 h-4" />) as any}
-                  onClick={() => applyMarkdownToolbarAction('heading')}
-                  disabled={editorDisabled}
-                  aria-label={t('option:notesSearch.toolbarHeadingTooltip', { defaultValue: 'Heading' })}
-                  data-testid="notes-toolbar-heading"
-                />
-              </Tooltip>
-              <Tooltip title={t('option:notesSearch.toolbarListTooltip', { defaultValue: 'List' })}>
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<ListIcon className="w-4 h-4" />) as any}
-                  onClick={() => applyMarkdownToolbarAction('list')}
-                  disabled={editorDisabled}
-                  aria-label={t('option:notesSearch.toolbarListTooltip', { defaultValue: 'List' })}
-                  data-testid="notes-toolbar-list"
-                />
-              </Tooltip>
-              <Tooltip title={t('option:notesSearch.toolbarLinkTooltip', { defaultValue: 'Link' })}>
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<LinkIcon className="w-4 h-4" />) as any}
-                  onClick={() => applyMarkdownToolbarAction('link')}
-                  disabled={editorDisabled}
-                  aria-label={t('option:notesSearch.toolbarLinkTooltip', { defaultValue: 'Link' })}
-                  data-testid="notes-toolbar-link"
-                />
-              </Tooltip>
-              <Tooltip title={t('option:notesSearch.toolbarAttachmentTooltip', { defaultValue: 'Attachment' })}>
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<PaperclipIcon className="w-4 h-4" />) as any}
-                  onClick={openAttachmentPicker}
-                  disabled={editorDisabled}
-                  aria-label={t('option:notesSearch.toolbarAttachmentTooltip', { defaultValue: 'Attachment' })}
-                  data-testid="notes-toolbar-attachment"
-                />
-              </Tooltip>
-              <Tooltip title={t('option:notesSearch.toolbarCodeTooltip', { defaultValue: 'Code' })}>
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<CodeIcon className="w-4 h-4" />) as any}
-                  onClick={() => applyMarkdownToolbarAction('code')}
-                  disabled={editorDisabled}
-                  aria-label={t('option:notesSearch.toolbarCodeTooltip', { defaultValue: 'Code' })}
-                  data-testid="notes-toolbar-code"
-                />
-              </Tooltip>
-              <span className="mx-1 h-4 w-px bg-border" aria-hidden="true" />
-              <Typography.Text
-                type="secondary"
-                className="text-[11px] mr-1 uppercase tracking-[0.08em]"
-              >
-                {t('option:notesSearch.assistLabel', {
-                  defaultValue: 'Assist'
-                })}
-              </Typography.Text>
-              <Tooltip
-                title={t('option:notesSearch.assistSummarizeTooltip', {
-                  defaultValue: 'Generate a concise summary draft'
-                })}
-              >
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<SparklesIcon className="w-4 h-4" />) as any}
-                  onClick={() => {
-                    void runAssistAction('summarize')
-                  }}
-                  disabled={editorDisabled || content.trim().length === 0}
-                  loading={assistLoadingAction === 'summarize'}
-                  data-testid="notes-assist-summarize"
-                >
-                  {t('option:notesSearch.assistSummarizeAction', {
-                    defaultValue: 'Summarize'
-                  })}
-                </Button>
-              </Tooltip>
-              <Tooltip
-                title={t('option:notesSearch.assistExpandOutlineTooltip', {
-                  defaultValue: 'Generate an expanded outline draft'
-                })}
-              >
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<SparklesIcon className="w-4 h-4" />) as any}
-                  onClick={() => {
-                    void runAssistAction('expand_outline')
-                  }}
-                  disabled={editorDisabled || content.trim().length === 0}
-                  loading={assistLoadingAction === 'expand_outline'}
-                  data-testid="notes-assist-expand-outline"
-                >
-                  {t('option:notesSearch.assistExpandOutlineAction', {
-                    defaultValue: 'Expand outline'
-                  })}
-                </Button>
-              </Tooltip>
-              <Tooltip
-                title={t('option:notesSearch.assistSuggestKeywordsTooltip', {
-                  defaultValue: 'Suggest keywords from note content'
-                })}
-              >
-                <Button
-                  size="small"
-                  type="text"
-                  icon={(<SparklesIcon className="w-4 h-4" />) as any}
-                  onClick={() => {
-                    void runAssistAction('suggest_keywords')
-                  }}
-                  disabled={editorDisabled || content.trim().length === 0}
-                  loading={assistLoadingAction === 'suggest_keywords'}
-                  data-testid="notes-assist-suggest-keywords"
-                >
-                  {t('option:notesSearch.assistSuggestKeywordsAction', {
-                    defaultValue: 'Suggest keywords'
-                  })}
-                </Button>
-              </Tooltip>
-              <input
-                ref={attachmentInputRef}
-                type="file"
-                multiple
-                className="hidden"
-                onChange={handleAttachmentInputChange}
-                data-testid="notes-attachment-input"
-              />
-            </div>
-          )}
-          {shouldShowToc && (
-            <div
-              className="mt-3 rounded-lg border border-border bg-surface2 p-2"
-              data-testid="notes-toc-panel"
-            >
-              <Typography.Text
-                type="secondary"
-                className="block text-[11px] uppercase tracking-[0.08em] text-text-muted"
-              >
-                {t('option:notesSearch.tocHeading', {
-                  defaultValue: 'Table of contents'
-                })}
-              </Typography.Text>
-              <div className="mt-2 space-y-1">
-                {tocEntries.map((entry) => (
-                  <button
-                    key={`toc-${entry.id}-${entry.offset}`}
-                    type="button"
-                    className="block w-full rounded px-2 py-1 text-left text-xs text-text hover:bg-surface3"
-                    style={{ paddingLeft: `${8 + (entry.level - 1) * 12}px` }}
-                    onClick={() => handleTocJump(entry)}
-                    data-testid={`notes-toc-item-${entry.id}`}
-                  >
-                    {entry.text}
-                  </button>
-                ))}
-              </div>
-            </div>
-          )}
-          <div className="mt-2 flex-1 min-h-0">
-            {editorMode === 'preview' ? (
-              content.trim().length > 0 ? (
-                <div className="h-full flex flex-col">
-                  <Typography.Text
-                    type="secondary"
-                    className="block text-[11px] mb-2 text-text-muted"
-                  >
-                    {t('option:notesSearch.previewTitle', {
-                      defaultValue: 'Preview (Markdown + LaTeX)'
-                    })}
-                  </Typography.Text>
-                  {usesLargePreviewGuardrails && !largePreviewReady ? (
-                    <div
-                      className="w-full flex-1 rounded-lg border border-border bg-surface2 p-4"
-                      role="status"
-                      aria-live="polite"
-                      data-testid="notes-large-preview-loading"
-                    >
-                      <div className="inline-flex items-center gap-2 text-sm text-text-muted">
-                        <Spin size="small" />
-                        <span>
-                          {t('option:notesSearch.largePreviewLoadingLabel', {
-                            defaultValue: 'Rendering preview for large note'
-                          })}
-                          {`: ${previewContent.length} chars`}
-                        </span>
-                      </div>
-                    </div>
-                  ) : (
-                    <div
-                      className="w-full flex-1 text-sm p-4 rounded-lg border border-border bg-surface2 overflow-auto"
-                      onClick={handlePreviewLinkClick}
-                      data-testid="notes-preview-surface"
-                    >
-                      <MarkdownPreview content={previewContent} size="sm" />
-                    </div>
-                  )}
-                </div>
-              ) : (
-                <Typography.Text
-                  type="secondary"
-                  className="block text-[11px] mt-1 text-text-muted"
-                >
-                  {t('option:notesSearch.emptyPreview', {
-                    defaultValue:
-                      'Start typing to see a live preview of your note.'
-                  })}
-                </Typography.Text>
-              )
-            ) : editorMode === 'split' ? (
-              <div className="grid h-full min-h-0 grid-cols-1 gap-3 lg:grid-cols-2">
-                <div className="flex min-h-0 flex-col">
-                  <Typography.Text
-                    type="secondary"
-                    className="block text-[11px] mb-2 text-text-muted"
-                  >
-                    {t('option:notesSearch.editModeLabel', {
-                      defaultValue: 'Edit'
-                    })}
-                  </Typography.Text>
-                  {editorInputMode === 'wysiwyg' ? (
-                    <div
-                      ref={richEditorRef}
-                      role="textbox"
-                      aria-multiline="true"
-                      contentEditable={!editorDisabled}
-                      suppressContentEditableWarning
-                      className="w-full min-h-[220px] text-sm p-4 rounded-lg border border-border bg-surface2 text-text overflow-auto leading-relaxed focus:outline-none focus:ring-2 focus:ring-focus"
-                      onInput={handleWysiwygInput}
-                      onPaste={handleWysiwygPaste}
-                      onBlur={() => setEditorCursorIndex(null)}
-                      aria-label={t('option:notesSearch.editorAriaLabel', {
-                        defaultValue: 'Note content'
-                      })}
-                      data-testid="notes-wysiwyg-editor"
-                      dangerouslySetInnerHTML={{ __html: wysiwygHtml }}
-                    />
-                  ) : (
-                    <>
-                      <textarea
-                        ref={contentTextareaRef}
-                        className="w-full min-h-[220px] text-sm p-4 rounded-lg border border-border bg-surface2 text-text resize-none leading-relaxed focus:outline-none focus:ring-2 focus:ring-focus"
-                        value={content}
-                        onChange={handleEditorChange}
-                        onKeyDown={handleEditorKeyDown}
-                        onSelect={handleEditorSelectionUpdate}
-                        onClick={handleEditorSelectionUpdate}
-                        onKeyUp={handleEditorSelectionUpdate}
-                        onFocus={handleEditorSelectionUpdate}
-                        onBlur={() => setEditorCursorIndex(null)}
-                        placeholder={t('option:notesSearch.editorPlaceholder', {
-                          defaultValue: 'Write your note here... (Markdown supported)'
-                        })}
-                        readOnly={editorDisabled}
-                        aria-label={t('option:notesSearch.editorAriaLabel', {
-                          defaultValue: 'Note content'
-                        })}
-                      />
-                      {activeWikilinkQuery && wikilinkSuggestions.length > 0 && (
-                        <div
-                          className="mt-2 rounded-lg border border-border bg-surface p-1"
-                          role="listbox"
-                          aria-label={t('option:notesSearch.wikilinkSuggestionsLabel', {
-                            defaultValue: 'Wikilink suggestions'
-                          })}
-                          data-testid="notes-wikilink-suggestions"
-                        >
-                          {wikilinkSuggestions.map((candidate, index) => {
-                            const duplicateCount =
-                              wikilinkSuggestionDisplayCounts.get(candidate.title.toLowerCase()) || 0
-                            const label =
-                              duplicateCount > 1 ? `${candidate.title} (${candidate.id})` : candidate.title
-                            return (
-                              <button
-                                key={`${candidate.id}-${candidate.title}`}
-                                type="button"
-                                className={`block w-full rounded px-2 py-1 text-left text-xs ${
-                                  index === wikilinkSelectionIndex
-                                    ? 'bg-surface2 text-text'
-                                    : 'text-text-muted hover:bg-surface2 hover:text-text'
-                                }`}
-                                aria-selected={index === wikilinkSelectionIndex}
-                                onMouseDown={(event) => {
-                                  event.preventDefault()
-                                  applyWikilinkSuggestion(candidate)
-                                }}
-                                data-testid={`notes-wikilink-suggestion-${candidate.id.replace(/[^a-z0-9_-]/gi, '_')}`}
-                              >
-                                {label}
-                              </button>
-                            )
-                          })}
-                        </div>
-                      )}
-                    </>
-                  )}
-                  <Typography.Text
-                    type="secondary"
-                    className="block text-[11px] mt-1 text-text-muted"
-                  >
-                    {editorInputMode === 'wysiwyg'
-                      ? t('option:notesSearch.wysiwygSupportHint', {
-                          defaultValue: 'WYSIWYG mode keeps markdown structure while you edit.'
-                        })
-                      : t('option:notesSearch.editorSupportHint', {
-                          defaultValue: 'Markdown + LaTeX supported'
-                        })}
-                  </Typography.Text>
-                </div>
-                <div className="flex min-h-0 flex-col">
-                  {content.trim().length > 0 ? (
-                    <>
-                      <Typography.Text
-                        type="secondary"
-                        className="block text-[11px] mb-2 text-text-muted"
-                      >
-                        {t('option:notesSearch.previewTitle', {
-                          defaultValue: 'Preview (Markdown + LaTeX)'
-                        })}
-                      </Typography.Text>
-                      {usesLargePreviewGuardrails && !largePreviewReady ? (
-                        <div
-                          className="w-full flex-1 rounded-lg border border-border bg-surface2 p-4"
-                          role="status"
-                          aria-live="polite"
-                          data-testid="notes-large-preview-loading"
-                        >
-                          <div className="inline-flex items-center gap-2 text-sm text-text-muted">
-                            <Spin size="small" />
-                            <span>
-                              {t('option:notesSearch.largePreviewLoadingLabel', {
-                                defaultValue: 'Rendering preview for large note'
-                              })}
-                              {`: ${previewContent.length} chars`}
-                            </span>
-                          </div>
-                        </div>
-                      ) : (
-                        <div
-                          className="w-full flex-1 text-sm p-4 rounded-lg border border-border bg-surface2 overflow-auto"
-                          onClick={handlePreviewLinkClick}
-                          data-testid="notes-split-preview-surface"
-                        >
-                          <MarkdownPreview content={previewContent} size="sm" />
-                        </div>
-                      )}
-                    </>
-                  ) : (
-                    <Typography.Text
-                      type="secondary"
-                      className="block text-[11px] mt-1 text-text-muted"
-                    >
-                      {t('option:notesSearch.emptyPreview', {
-                        defaultValue:
-                          'Start typing to see a live preview of your note.'
-                      })}
-                    </Typography.Text>
-                  )}
-                </div>
-              </div>
-            ) : (
-              <div className="flex h-full min-h-0 flex-col">
-                {editorInputMode === 'wysiwyg' ? (
-                  <div
-                    ref={richEditorRef}
-                    role="textbox"
-                    aria-multiline="true"
-                    contentEditable={!editorDisabled}
-                    suppressContentEditableWarning
-                    className="w-full min-h-[280px] text-sm p-4 rounded-lg border border-border bg-surface2 text-text overflow-auto leading-relaxed focus:outline-none focus:ring-2 focus:ring-focus"
-                    onInput={handleWysiwygInput}
-                    onPaste={handleWysiwygPaste}
-                    onBlur={() => setEditorCursorIndex(null)}
-                    aria-label={t('option:notesSearch.editorAriaLabel', {
-                      defaultValue: 'Note content'
-                    })}
-                    data-testid="notes-wysiwyg-editor"
-                    dangerouslySetInnerHTML={{ __html: wysiwygHtml }}
-                  />
-                ) : (
-                  <>
-                    <textarea
-                      ref={contentTextareaRef}
-                      className="w-full min-h-[280px] text-sm p-4 rounded-lg border border-border bg-surface2 text-text resize-none leading-relaxed focus:outline-none focus:ring-2 focus:ring-focus"
-                      value={content}
-                      onChange={handleEditorChange}
-                      onKeyDown={handleEditorKeyDown}
-                      onSelect={handleEditorSelectionUpdate}
-                      onClick={handleEditorSelectionUpdate}
-                      onKeyUp={handleEditorSelectionUpdate}
-                      onFocus={handleEditorSelectionUpdate}
-                      onBlur={() => setEditorCursorIndex(null)}
-                      placeholder={t('option:notesSearch.editorPlaceholder', {
-                        defaultValue: 'Write your note here... (Markdown supported)'
-                      })}
-                      readOnly={editorDisabled}
-                      aria-label={t('option:notesSearch.editorAriaLabel', {
-                        defaultValue: 'Note content'
-                      })}
-                    />
-                    {activeWikilinkQuery && wikilinkSuggestions.length > 0 && (
-                      <div
-                        className="mt-2 rounded-lg border border-border bg-surface p-1"
-                        role="listbox"
-                        aria-label={t('option:notesSearch.wikilinkSuggestionsLabel', {
-                          defaultValue: 'Wikilink suggestions'
-                        })}
-                        data-testid="notes-wikilink-suggestions"
-                      >
-                        {wikilinkSuggestions.map((candidate, index) => {
-                          const duplicateCount =
-                            wikilinkSuggestionDisplayCounts.get(candidate.title.toLowerCase()) || 0
-                          const label =
-                            duplicateCount > 1 ? `${candidate.title} (${candidate.id})` : candidate.title
-                          return (
-                            <button
-                              key={`${candidate.id}-${candidate.title}`}
-                              type="button"
-                              className={`block w-full rounded px-2 py-1 text-left text-xs ${
-                                index === wikilinkSelectionIndex
-                                  ? 'bg-surface2 text-text'
-                                  : 'text-text-muted hover:bg-surface2 hover:text-text'
-                              }`}
-                              aria-selected={index === wikilinkSelectionIndex}
-                              onMouseDown={(event) => {
-                                event.preventDefault()
-                                applyWikilinkSuggestion(candidate)
-                              }}
-                              data-testid={`notes-wikilink-suggestion-${candidate.id.replace(/[^a-z0-9_-]/gi, '_')}`}
-                            >
-                              {label}
-                            </button>
-                          )
-                        })}
-                      </div>
-                    )}
-                  </>
-                )}
-                <Typography.Text
-                  type="secondary"
-                  className="block text-[11px] mt-1 text-text-muted"
-                >
-                  {editorInputMode === 'wysiwyg'
-                    ? t('option:notesSearch.wysiwygSupportHint', {
-                        defaultValue: 'WYSIWYG mode keeps markdown structure while you edit.'
-                      })
-                    : t('option:notesSearch.editorSupportHint', {
-                        defaultValue: 'Markdown + LaTeX supported'
-                      })}
-                </Typography.Text>
-              </div>
-            )}
-          </div>
-          <div className="mt-2 border-t border-border pt-2">
-            <Typography.Text
-              type="secondary"
-              className="text-[11px] text-text-muted"
-              data-testid="notes-editor-metrics"
-            >
-              {metricSummaryText}
-            </Typography.Text>
-            <Typography.Text
-              type="secondary"
-              className="block text-[11px] text-text-muted mt-1"
-              data-testid="notes-editor-revision-meta"
-            >
-              {revisionSummaryText}
-            </Typography.Text>
-            <Typography.Text
-              type="secondary"
-              className="block text-[11px] text-text-muted mt-1"
-              data-testid="notes-editor-provenance"
-            >
-              {provenanceSummaryText}
-            </Typography.Text>
-            {queuedOfflineDraftCount > 0 && (
-              <Typography.Text
-                type="secondary"
-                className="block text-[11px] text-text-muted mt-1"
-                data-testid="notes-editor-offline-queue-meta"
-              >
-                {t('option:notesSearch.offlineQueueFooterMeta', {
-                  defaultValue: 'Queued offline drafts: {{count}}',
-                  count: queuedOfflineDraftCount
-                })}
-              </Typography.Text>
-            )}
-          </div>
-        </div>
-      </section>
+      <NotesEditorPane
+        isMobileViewport={isMobileViewport}
+        setMobileSidebarOpen={setMobileSidebarOpen}
+        selectedId={selectedId}
+        title={title}
+        content={content}
+        editorDisabled={editorDisabled}
+        isDirty={isDirty}
+        isOnline={isOnline}
+        loadingDetail={loadingDetail}
+        saving={saving}
+        backlinkConversationId={backlinkConversationId}
+        backlinkConversationLabel={backlinkConversationLabel}
+        backlinkMessageId={backlinkMessageId}
+        noteRelations={noteRelations}
+        noteNeighborsLoading={noteNeighborsLoading}
+        noteNeighborsError={noteNeighborsError}
+        selectedNotePinned={selectedNotePinned}
+        editorMode={editorMode}
+        editorInputMode={editorInputMode}
+        openingLinkedChat={openingLinkedChat}
+        editorKeywords={editorKeywords}
+        keywordOptions={keywordOptions}
+        saveIndicator={saveIndicator}
+        saveIndicatorText={saveIndicatorText}
+        offlineStatusText={offlineStatusText}
+        currentOfflineDraft={currentOfflineDraft}
+        remoteVersionInfo={remoteVersionInfo}
+        monitoringNotice={monitoringNotice}
+        monitoringNoticeClasses={monitoringNoticeClasses}
+        titleSuggestionLoading={titleSuggestionLoading}
+        canSwitchTitleStrategy={canSwitchTitleStrategy}
+        effectiveTitleSuggestStrategy={effectiveTitleSuggestStrategy}
+        titleStrategyOptions={titleStrategyOptions}
+        setTitleSuggestStrategy={setTitleSuggestStrategy}
+        manualLinkTargetId={manualLinkTargetId}
+        setManualLinkTargetId={setManualLinkTargetId}
+        manualLinkSaving={manualLinkSaving}
+        manualLinkOptions={manualLinkOptions}
+        manualLinkDeletingEdgeId={manualLinkDeletingEdgeId}
+        assistLoadingAction={assistLoadingAction}
+        shouldShowToc={shouldShowToc}
+        tocEntries={tocEntries}
+        previewContent={previewContent}
+        usesLargePreviewGuardrails={usesLargePreviewGuardrails}
+        largePreviewReady={largePreviewReady}
+        wysiwygHtml={wysiwygHtml}
+        activeWikilinkQuery={activeWikilinkQuery}
+        wikilinkSuggestions={wikilinkSuggestions}
+        wikilinkSuggestionDisplayCounts={wikilinkSuggestionDisplayCounts}
+        wikilinkSelectionIndex={wikilinkSelectionIndex}
+        metricSummaryText={metricSummaryText}
+        revisionSummaryText={revisionSummaryText}
+        provenanceSummaryText={provenanceSummaryText}
+        queuedOfflineDraftCount={queuedOfflineDraftCount}
+        titleInputRef={titleInputRef}
+        contentTextareaRef={contentTextareaRef}
+        richEditorRef={richEditorRef}
+        attachmentInputRef={attachmentInputRef}
+        setTitle={setTitle}
+        setIsDirty={setIsDirty}
+        setSaveIndicator={setSaveIndicator}
+        setMonitoringNotice={setMonitoringNotice}
+        setEditorMode={setEditorMode}
+        setEditorKeywords={setEditorKeywords}
+        setEditorCursorIndex={setEditorCursorIndex}
+        setShortcutHelpOpen={setShortcutHelpOpen}
+        markManualEdit={markManualEdit}
+        suggestTitle={suggestTitle}
+        openLinkedConversation={openLinkedConversation}
+        openLinkedSource={openLinkedSource}
+        handleNewNote={handleNewNote}
+        duplicateSelectedNote={duplicateSelectedNote}
+        toggleNotePinned={toggleNotePinned}
+        copySelected={copySelected}
+        handleGenerateFlashcardsFromNote={handleGenerateFlashcardsFromNote}
+        exportSelected={exportSelected}
+        saveNote={saveNote}
+        deleteNote={deleteNote}
+        handleSelectNote={handleSelectNote}
+        openGraphModal={openGraphModal}
+        createManualLink={createManualLink}
+        removeManualLink={removeManualLink}
+        debouncedLoadKeywordSuggestions={debouncedLoadKeywordSuggestions}
+        renderKeywordLabelWithFrequency={renderKeywordLabelWithFrequency}
+        handleEditorInputModeChange={handleEditorInputModeChange}
+        applyMarkdownToolbarAction={applyMarkdownToolbarAction}
+        openAttachmentPicker={openAttachmentPicker}
+        handleAttachmentInputChange={handleAttachmentInputChange}
+        runAssistAction={runAssistAction}
+        handleTocJump={handleTocJump}
+        handlePreviewLinkClick={handlePreviewLinkClick}
+        handleWysiwygInput={handleWysiwygInput}
+        handleWysiwygPaste={handleWysiwygPaste}
+        handleEditorChange={handleEditorChange}
+        handleEditorKeyDown={handleEditorKeyDown}
+        handleEditorSelectionUpdate={handleEditorSelectionUpdate}
+        applyWikilinkSuggestion={applyWikilinkSuggestion}
+      />
       <Modal
         open={keywordSuggestionOptions.length > 0}
         onCancel={closeKeywordSuggestionModal}
